chore: move GoogleAnalytics out to global component
